Write an explanation what the urlencoded option is used for in the middleware
// The extended option allows to choose between parsing the URL-encoded dat
a with the querystring library (when false) or the qs library (when true).
The "extended" syntax allows for rich objects and arrays to be encoded into
the URL-encoded format, allowing for a JSON-like experience with URL-encod
ed. For more information, please see the qs library. /*
foo[bar]=baz', that's a qs type so nested json can happen (rare in the wild
) /?mario=five&you=two is a querystring (super common) */